requirements.txt
torch>=2.6,<2.8
transformers>=4.57.0,<4.59
accelerate>=1.5.0,<2.0
huggingface_hub>=0.34.0,<1.0
peft>=0.15.0
datasets>=3.3.0
bitsandbytes>=0.45.0
safetensors>=0.5.0
wandb>=0.18
pyyaml>=6.0
python-dotenv>=1.0
tqdm>=4.66
sentencepiece>=0.2.0
